% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareMins.R
\name{compareMins}
\alias{compareMins}
\title{Utility function to identify minimum values}
\usage{
compareMins(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A logical vector of length input vector minus 1, corresponding to whether an
element of the input vector equals the last element of the input vector.
}
\description{
Given a vector where the last element is the minimum, identifies which elements in that
vector match the last element.
}
\details{
Simple utility function, used in the competitive exclusion simulations.
}
\examples{
#create a basic input vector
temp <- c(1,2,3,4,5,6,1)

#here is an example of the compareMins function
compareMins(temp)
}
\references{
Miller, Trisos and Farine.
}

