% Generated by roxygen2 (4.0.2): do not edit by hand
\name{defineMetrics}
\alias{defineMetrics}
\title{Output all metrics as a list of named functions}
\usage{
defineMetrics()
}
\value{
A list of named functions
}
\description{
Creates a list of named functions, each of which accept a metrics.input object
}
\details{
All of the functions we calculated for our manuscript are included in this
function. To add additional metric functions, they need to be defined and included
here. The function needs to be included with a name, and it must accept a metrics.input
as input. If the function needs additional elements not included in that input, then
the prepData function must also be revised.
}
\examples{
library(geiger)
library(picante)

defineMetrics()
}
\references{
Miller, Trisos and Farine.
}

