% Generated by roxygen2 (4.0.2): do not edit by hand
\name{defineNulls}
\alias{defineNulls}
\title{Output all null models as a list of named functions}
\usage{
defineNulls()
}
\value{
A list of named functions
}
\description{
Creates a list of named functions, each of which accept a nulls.input object
}
\details{
All of the null models we calculated for our manuscript are included in this
function. To add additional functions, they need to be defined and included
here. The function needs to be included with a name, and it must accept a nulls.input
as input. If the function needs additional elements not included in that input, then
the prepNulls function must also be revised.
}
\examples{
library(geiger)
library(picante)

defineNulls()
}
\references{
Miller, Trisos and Farine.
}

