% Generated by roxygen2 (4.0.2): do not edit by hand
\name{defineSimulations}
\alias{defineSimulations}
\title{Output all spatial simulations as a list of named functions}
\usage{
defineSimulations()
}
\value{
A list of named functions
}
\description{
Creates a list of named functions, each of which accept a simulations.input object
}
\details{
All of the spatial simulations we calculated for our manuscript are included
in this function. To add additional functions, they need to be defined and included
here. The function needs to be included with a name, and it must accept a
simulations.input object. If the function needs additional elements not included in
that input, then the prepSimulations function must also be revised.
}
\examples{
library(geiger)
library(picante)

defineSimulations()
}
\references{
Miller, Trisos and Farine.
}

