% Generated by roxygen2 (4.0.2): do not edit by hand
\name{errorChecker}
\alias{errorChecker}
\title{Wrapper for summarizing randomizations and testing significance of observed metrics}
\usage{
errorChecker(observed, reduced.randomizations, concat.by)
}
\arguments{
\item{observed}{Data frame of observed metric scores, such as from observedMetrics()}

\item{reduced.randomizations}{List of random, reduced results, such as those from
reduceRandomizations()}

\item{concat.by}{Whether to concatenate the randomizations by richness or quadrat}
}
\value{
A list of lists of data frames.
}
\description{
Given a data frame of observed scores and a list of randomizations based on different
null models, returns a list of data frames summarizing the significance of observed
scores both at the single quadrat and the entire arena level.
}
\details{
This function wraps a number of smaller functions into a useful type I and II
error checker. It takes a reduced list of randomizations such as those reduced from
metricsNnulls with reduceRandomizations, summarizes the mean,
SD, and CI of each metric plus null model either at the richness or quadrat level,
then compares the observed metric scores to those summarized scores. It return a list
with two elements. The first is a list of data frames, where each corresponds to the
standardized effect scores of the observed metrics for a given null model. The second
is a list of data frames, where each corresponds to whether each observed metric
deviates beyond expectations.
}
\examples{
library(dplyr)
library(geiger)
library(picante)

#simulate tree with birth-death process
tree <- sim.bdtree(b=0.1, d=0, stop="taxa", n=50)

#simulate a log normal abundance distribution
sim.abundances <- round(rlnorm(5000, meanlog=2, sdlog=1)) + 1

#simulate a community of varying richness
cdm <- simulateComm(tree, min.rich=10, max.rich=25, abundances=sim.abundances)

#run the metrics and nulls combo function
rawResults <- metricsNnulls(tree, cdm, randomizations=3)

#summarize the results
results <- reduceRandomizations(rawResults)

#calculate the observed metrics from the input CDM
observed <- observedMetrics(tree, cdm)

test <- errorChecker(observed, results, "richness")
}
\references{
Miller, Trisos and Farine.
}

