% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evolveTraits}
\alias{evolveTraits}
\title{Evolve two traits up a tree}
\usage{
evolveTraits(tree, sigma = 0.1)
}
\arguments{
\item{tree}{A phylogeny}
}
\value{
A list where the first object is a phylogeny with the desired number of species
and the second object is a matrix of trait values for those species.
}
\description{
Given a phylogeny, will generate associated trait data for two traits following a
Brownian motion evolution model.
}
\details{
Evolves two traits up phylogeny with Brownian motion evolution process. Sigma
is currently set to 0.1 as default.
}
\examples{
library(geiger)

tree <- sim.bdtree(b=0.1, d=0, stop="taxa", n=50)

results <- evolveTraits(tree)
}
\references{
Miller, Trisos and Farine.
}

