% Generated by roxygen2 (4.0.2): do not edit by hand
\name{failed}
\alias{failed}
\title{Identify failed runs}
\usage{
failed(single.iteration)
}
\arguments{
\item{single.iteration}{The results of a single iteration of multiLinker}
}
\description{
This identifies spatial sim/null model runs that failed
}
\details{
It is possible that a given null model, e.g. regional, failed on a given run.
If this is the case, one can identify it with this function and remove it from the
results of that iteration. This prevents errors further down the line when summarizing
overall results.
}
\references{
Miller, Trisos and Farine.
}

