% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filteringArena}
\alias{filteringArena}
\title{Simulate a community assembled according to habitat filtering}
\usage{
filteringArena(simulations.input)
}
\arguments{
\item{simulations.input}{A prepared simulations.input object from prepSimulations}
}
\value{
A list of 3 elements: the original input regional
abundance vector, the new spatial arena, and the dimensions of that arena.
}
\description{
Given a simulations.input object, will create an arena settled according to habitat
filtering rules (and parameters defined by prepSimulations).
}
\details{
This simulation is sensitive to parameter specifications. Because there are
no "forces" deterring individuals from settling near each other, species clump near
their spatial preferences.
}
\examples{
library(geiger)

tree <- sim.bdtree(b=0.1, d=0, stop="taxa", n=50)

prepped <- prepSimulations(tree, arena.length=300, mean.log.individuals=4,
length.parameter=5000, sd.parameter=50, max.distance=20, proportion.killed=0.2,
competition.iterations=2)

test <- filteringArena(prepped)
}
\references{
Miller, Trisos and Farine.
}

