% Generated by roxygen2 (4.0.2): do not edit by hand
\name{linker}
\alias{linker}
\title{Run spatial simulations, null and metric calculations to test metric + null performance}
\usage{
linker(tree, arena.length, mean.log.individuals, length.parameter, sd.parameter,
  max.distance, proportion.killed, competition.iterations, no.quadrats,
  quadrat.length, concat.by, randomizations, cores)
}
\arguments{
\item{tree}{Phylo object}

\item{arena.length}{A numeric, specifying the length of a single side of the arena}

\item{mean.log.individuals}{Mean log of abundance vector from which species abundances
will be drawn}

\item{length.parameter}{Length of vector from which species' locations are drawn. Large
values of this parameter dramatically decrease the speed of the function but result in
nicer looking communities}

\item{sd.parameter}{Standard deviation of vector from which species' locations are
drawn}

\item{max.distance}{The geographic distance within which neighboring
indivduals should be considered to influence the individual in question}

\item{proportion.killed}{The percent of individuals in the total arena that should be
considered (as a proportion, e.g. 0.5 = half)}

\item{competition.iterations}{Number of generations over which to run competition
simulations}

\item{no.quadrats}{Number of quadrats to place}

\item{quadrat.length}{Length of one side of desired quadrat}

\item{concat.by}{Whether to concatenate the randomizations by richness or quadrat}

\item{randomizations}{The number of randomized CDMs, per null, to generate. These are
used to compare the significance of the observed metric scores.}

\item{cores}{The number of cores to be used for parallel processing.}
}
\value{
A list of lists of data frames. The first level of the output has one element
for each simulation. The second level has one element for each null model. Each of
these elements is a list of two data frames, one that summarizes the quadrat-level
significance and another and arena-level significance.
}
\description{
This function wraps a number of wrapper functions into one big metric + null
tester function. Only a single test is performed, with results saved into memory.
}
\details{
This function wraps a number of other wrapper functions into
one big metric + null performance tester function. Only a single test is performed,
with results saved into memory. To perform multiple complete tests, use the
multiLinker function, which saves results to file.
}
\examples{
library(dplyr)
library(geiger)
library(picante)

#simulate tree with birth-death process
tree <- sim.bdtree(b=0.1, d=0, stop="taxa", n=50)

system.time(test <- linker(tree=tree, arena.length=300, mean.log.individuals=2,
	length.parameter=5000, sd.parameter=50, max.distance=30, proportion.killed=0.2,
	competition.iterations=3, no.quadrats=15, quadrat.length=30, concat.by="richness",
	randomizations=3, cores=3))
}
\references{
Miller, Trisos and Farine.
}

