% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phyloField.R
\name{phyloField}
\alias{phyloField}
\title{Calculate a species' phylogenetic field}
\usage{
phyloField(tree, picante.cdm, metric)
}
\arguments{
\item{tree}{Phylo object}

\item{picante.cdm}{A picante-style community data matrix with sites as rows, and
species as columns}

\item{metric}{Phylogenetic metric of choice (see details)}
}
\value{
Named vector of species' phylogenetic fields.
}
\description{
Calculate the phylogenetic relatedness of a species to those it occurs with.
}
\details{
Currently this is only programmed to use either non-abundance-weighted mean
pairwise or interspecific abundance-weighted mean pairwise phylogenetic distance.
The function could be improved by tapping into any of the phylogenetic metrics defined
in defineMetrics.
}
\examples{
#simulate tree with birth-death process
tree <- geiger::sim.bdtree(b=0.1, d=0, stop="taxa", n=50)

#simulate log-normal abundances
sim.abundances <- round(rlnorm(5000, meanlog=2, sdlog=1)) + 1

#simulate a community data matrix with these inputs
cdm <- simulateComm(tree, richness.vector=10:25, abundances=sim.abundances)

#example phylogenetic field calculations
exampleField <- phyloField(tree=tree, picante.cdm=cdm, metric="naw.mpd")
}
\references{
Miller, Wagner, Harmon & Ricklefs. In review. Radiating despite a lack of
character: closely related, morphologically similar, co-occurring honeyeaters have
diverged ecologically.
}

