% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverall.R
\name{plotOverall}
\alias{plotOverall}
\title{Overall per simulation-null-metric plot test}
\usage{
plotOverall(simulation.list)
}
\arguments{
\item{simulation.list}{A summarized results list such as one output from
reduceResults(). See examples.}
}
\value{
A data frame summarizing the total number of plots tested and how many of
these deviated above (significantly overdispersed) or below (significantly clustered)
the 95\% CI for each simulation, null, metric combination.
}
\description{
This function provides one of many ways of summarizing and considering simulation
results.
}
\details{
This function provides one way of summarizing and considering simulation
results. It takes as input a vector of 0s, 1s and 2s (corresponding to within, less,
and greater than the 95\% CIs, respectively) for all plots
from a given simulation-null-metric combination, and determines how many plots
overall deviated beyond expectations. A number of utility functions used for that goal
are also defined but not exported in this function.
}
\examples{
#not run
#results <- readIn()
#summ <- reduceResults(results)
#test <- plotOverall(summ$plot)
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2016. Phylogenetic community
structure metrics and null models: a review with new methods and software.
Ecography DOI: 10.1111/ecog.02070
}
