% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quadratContents.R
\name{quadratContents}
\alias{quadratContents}
\title{Identify individuals contained within a quadrat}
\usage{
quadratContents(arena, quadratPlacer.results)
}
\arguments{
\item{arena}{Data frame of three columns: "individuals", "X", and "Y"}

\item{quadratPlacer.results}{The results of a call to quadratPlacer. A list with two
elementers. The first is a matrix of X Y coordinates of quadrats. The second is a
symmetrical distance matrix summarizing the distances among these quadrats.}
}
\value{
A matrix with species as rows and quadrats as columns.
}
\description{
Given a spatially explicit data frame of individual locations in a simulated arena,
and the bounds of a series of quadrats, identifies the contents of each quadrat.
}
\details{
Takes a data frame like that returned from filteringArena(), and a matrix
like that returned from quadratPlacer(), and returns the resulting community data
matrix such as might be generated by someone surveying a forest plot. Quadrats with < 2
species are excluded from the CDM.
}
\examples{
library(colorRamps)

tree <- sim.bdtree(b=0.1, d=0, stop="taxa", n=50)

temp <- evolveTraits(tree)

phydistmatrix <- cophenetic(temp[[1]])

#define a color for each species
cols <- blue2green2red(nrow(phydistmatrix))

#prep the data for the simulation
prepped <- prepSimulations(tree, arena.length=300, mean.log.individuals=2,
length.parameter=5000, sd.parameter=50, max.distance=20, proportion.killed=0.2,
competition.iterations=3)

singleArena <- filteringArena(prepped)

#plot the arena. don't close the window
plot(singleArena$arena$X, singleArena$arena$Y, pch=20, cex=0.5, xlim=c(0,300),
	ylim=c(0,300), col=cols[singleArena$arena$individuals])

boundResults <- quadratPlacer(no.quadrats=10, arena.length=300, quadrat.length=50)

quadratPlotter(boundResults$quadrat.bounds)

#return a CDM in picante format
cdm <- quadratContents(singleArena$arena, boundResults)
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

