% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readIn}
\alias{readIn}
\title{Read in the results of multiple metric/null/simulation tests}
\usage{
readIn(working.directory)
}
\arguments{
\item{working.directory}{Optional character string specifying the working directory.
If missing, the current working directory will be used.}
}
\value{
A list of simulation results. Each element in the list consists of a single
output from multiLinker().
}
\description{
This function is used to read all .RDS files from the working directory into a
single list. This list is then summarized with additional functions.
}
\details{
This function reads all .RDS files from the working directory (or another
specified directory) into a single list.
}
\references{
Miller, Trisos and Farine.
}

