% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sesIndiv}
\alias{sesIndiv}
\title{Summary statistics of SES results}
\usage{
sesIndiv(raw.results)
}
\arguments{
\item{raw.results}{A list of lists of lists of dataframes; the results of a call to
readIn.}
}
\description{
Summarizes individual iteration performance of each sim/null/metric combination across
all iterations
}
\details{
This function takes a raw list of results from multiple iterations from
multiLinker, and runs sesSingle across each one. It then summarizes the results of
those single runs as the number of sim/null/metrics that deviated beyond expectations
and the number that were within expectations.
}
\examples{
#not run
#results <- readIn()
#summ <- sesIndiv(results)
}
\references{
Miller, Trisos and Farine.
}

