% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sesOverall}
\alias{sesOverall}
\title{Overall per simulation-null-metric SES test}
\usage{
sesOverall(simulation.list)
}
\arguments{
\item{simulation.list}{A summarized results list such as one output from
reduceResults(). See examples.}
}
\description{
This function provides one of many ways of summarizing and considering simulation
results.
}
\details{
This function provides one way of summarizing and considering simulation
results. It takes as input a vector of all standardized effect sizes for all quadrats
from a given simulation-null-metric combination, and calculates the mean of the vector
and whether it differs significantly from a mean of zero.
}
\examples{
#not run
#results <- readIn()
#summ <- reduceResults(results)
#test <- sesOverall(summ$ses)
}
\references{
Miller, Trisos and Farine.
}

