% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulateComm}
\alias{simulateComm}
\title{Generate a simulated community data matrix}
\usage{
simulateComm(tree, richness.vector, abundances)
}
\arguments{
\item{tree}{Phylo object}

\item{richness.vector}{Vector of desired species richness, one for each desired quadrat}

\item{abundances}{A vector of potential abundances, e.g. a log-normal distribution}
}
\value{
A community data matrix (a data frame) with species as columns and sites as
rows.
}
\description{
Given a phylo object, desired min and max species richnesses, and a vector of potential
species abundances, will generate a community data matrix with these characteristics.
}
\details{
There is currently no implementation to control the frequency with which a
given species is selected.
}
\examples{
library(geiger)
library(picante)

#simulate tree with birth-death process
tree <- sim.bdtree(b=0.1, d=0, stop="taxa", n=50)

sim.abundances <- round(rlnorm(5000, meanlog=2, sdlog=1)) + 1

cdm <- simulateComm(tree, richness.vector=10:25, abundances=sim.abundances)
}
\references{
Miller, Trisos and Farine.
}

