% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tWrap.R
\name{tWrap}
\alias{tWrap}
\title{Simple wrapper for t-test}
\usage{
tWrap(vect, mu = 0)
}
\arguments{
\item{vect}{A numeric vector}

\item{mu}{Proposed true mean. Default is 0.}
}
\value{
A one-row matrix, where the first column is the mean of the input vector, and
the second is the p.value as to whether the vector differs from the proposed true mean.
}
\description{
Given a numeric vector and a proposed true mean value, return the mean of the vector
and the p.value as to whether it differs from the proposed mean.
}
\details{
Given a vector and a proposed true mean value (default is mu=0), return the
mean of the vector and the p.value as to whether it differs from the proposed mean.
}
\examples{
a <- rnorm(100)
tWrap(a)
}
\references{
Miller, Trisos and Farine.
}

