% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tWrapApply.R
\name{tWrapApply}
\alias{tWrapApply}
\title{Apply wrapper for the tWrap function}
\usage{
tWrapApply(dataframe)
}
\arguments{
\item{dataframe}{A dataframe of numeric values}
}
\value{
A dataframe, with one row for each metric. The first column is the mean of the
vector of metric values, the second is the p.value of whether it differs from mu=0,
and the third is the name of the metric.
}
\description{
Applies tWrap over a dataframe of metric values, excluding richness.
}
\details{
Applies tWrap over a dataframe of metric values, excluding richness, which is
considered a metric in some previous functions. Note that there is currently no easy
way to pass arguments (like changing mu or whether the test is two- or one-tailed) down
to tWrap.
}
\examples{
a <- rnorm(n=100)
b <- rnorm(n=100, mean=100)
ex <- data.frame(a, b)
tWrapApply(ex)
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

