% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tWrapLApply.R
\name{tWrapLApply}
\alias{tWrapLApply}
\title{lapply wrapper for t-tests}
\usage{
tWrapLApply(null.list)
}
\arguments{
\item{null.list}{A list of dataframes, one per null model, of observed metric scores.}
}
\value{
A dataframe, with one row for each metric. The first column is the mean of the
vector of metric values, the second is the p.value of whether it differs from mu=0,
and the third is the name of the metric.
}
\description{
Just a utility function. lapplies t-tests over lists of data frames
}
\details{
lapplies tWrapApply over a list of dataframes. There is currently no easy
way to pass arguments (like changing mu or whether the test is two- or one-tailed) down
to tWrap.
}
\examples{
a <- rnorm(n=100)
b <- rnorm(n=100, mean=100)
ex <- data.frame(a, b)
test <- list("ex1"=ex, "ex2"=ex)
tWrapLApply(test)
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

