% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varLandscape.R
\name{varLandscape}
\alias{varLandscape}
\title{New spatial sim guts}
\usage{
varLandscape(cells, exponent = 1, seeds = 1)
}
\arguments{
\item{cells}{The number of cells to divide each side of the arena into.}

\item{exponent}{The exponent to which the distances will be raised. Default is 1.
values larger than 1 have the effect of , while values smaller than 1 have the effect
of .}

\item{seeds}{The number of "peaks" or trait "optima" that will be chosen in the
landscape. Default is 1.}
}
\value{
A matrix with more details
}
\description{
Longer description
}
\details{
More details
}
\references{
Miller, E. T. 2016. A new dispersal-informed null model for
community ecology shows strong performance. biorxiv.
}

