% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wilcoWrap.R
\name{wilcoWrap}
\alias{wilcoWrap}
\title{Simple wrapper for Wilcoxon signed rank test}
\usage{
wilcoWrap(vect, mu = 0, alternative)
}
\arguments{
\item{vect}{A numeric vector}

\item{mu}{Proposed true mean. Default is 0.}

\item{alternative}{Optional alternative hypothesis. Default is "two-sided". Use
"greater" for competition; "less" for habitat filtering.}
}
\value{
A one-row matrix, where the first column is the mean of the input vector, and
the second is the p.value as to whether the vector differs from the proposed true mean.
}
\description{
Given a numeric vector, a proposed true mean value, and an optional alternative
hypothesis, return the mean of the input vector and p.value.
}
\details{
Given a vector, a proposed true mean value (default is mu=0), and an
optional alternative hypothesis, return the mean of the vector and the p.value as to
whether it differs from the proposed mean. Using "greater" or "less" employ one-sided
tests when there is good reason for proposing the alternative hypothesis.
}
\examples{
a <- rnorm(100)
wilcoWrap(a)
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

