% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wilcoWrapApply.R
\name{wilcoWrapApply}
\alias{wilcoWrapApply}
\title{Apply wrapper for the wilcoWrap function}
\usage{
wilcoWrapApply(dataframe, alternative)
}
\arguments{
\item{dataframe}{A dataframe of numeric values}

\item{alternative}{Optional alternative hypothesis. Default is "two-sided". Use
"greater" for competition; "less" for habitat filtering.}
}
\value{
A dataframe, with one row for each metric. The first column is the mean of the
vector of metric values, the second is the p.value of whether it differs from mu=0,
and the third is the name of the metric.
}
\description{
Applies wilcoWrap over a dataframe of metric values, excluding richness.
}
\details{
Applies wilcoWrap over a dataframe of metric values, excluding richness, which
is considered a metric in some previous functions.
}
\examples{
a <- rnorm(n=100)
b <- rnorm(n=100, mean=100)
ex <- data.frame(a, b)
wilcoWrapApply(ex, alternative="two.sided")
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

