% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wilcoWrapLApply}
\alias{wilcoWrapLApply}
\title{lapply wrapper for the wilcoWrapApply function}
\usage{
wilcoWrapLApply(null.list, alternative)
}
\arguments{
\item{null.list}{A list of dataframes, one per null model, of observed metric scores.}
}
\value{
A dataframe, with one row for each metric. The first column is the mean of the
vector of metric values, the second is the p.value of whether it differs from mu=0,
and the third is the name of the metric.
}
\description{
lapplies wilcoWrapApply over a list of dataframes.
}
\details{
lapplies wilcoWrapApply over a list of dataframes.
}
\examples{
a <- rnorm(n=100)
b <- rnorm(n=100, mean=100)
ex <- data.frame(a, b)
test <- list("ex1"=ex, "ex2"=ex)
wilcoWrapLApply(test, alternative="two.sided")
}
\references{
Miller, Trisos and Farine.
}

