% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaLinker.R
\name{betaLinker}
\alias{betaLinker}
\title{Run spatial simulations, null and beta metric calculations}
\usage{
betaLinker(no.taxa, arena.length, mean.log.individuals, length.parameter,
  sd.parameter, max.distance, proportion.killed, competition.iterations,
  no.plots, plot.length, randomizations, cores, simulations, nulls,
  metrics)
}
\arguments{
\item{no.taxa}{The desired number of species in the input phylogeny}

\item{arena.length}{A numeric, specifying the length of a single side of the arena}

\item{mean.log.individuals}{Mean log of abundance vector from which species abundances
will be drawn}

\item{length.parameter}{Length of vector from which species' locations are drawn. Large
values of this parameter dramatically decrease the speed of the function but result in
nicer looking communities}

\item{sd.parameter}{Standard deviation of vector from which species' locations are 
drawn}

\item{max.distance}{The geographic distance within which neighboring
individuals should be considered to influence the individual in question}

\item{proportion.killed}{The percent of individuals in the total arena that should be
considered (as a proportion, e.g. 0.5 = half)}

\item{competition.iterations}{Number of generations over which to run competition 
simulations}

\item{no.plots}{Number of plots to place}

\item{plot.length}{Length of one side of desired plot}

\item{randomizations}{The number of randomized CDMs, per null, to generate. These are
used to compare the significance of the observed metric scores.}

\item{cores}{The number of cores to be used for parallel processing.}

\item{simulations}{Optional. If not provided, defines the simulations as all of those
in defineSimulations. If only a subset of those simulations is desired, then
simulations should take the form of a character vector corresponding to named functions
from defineSimulations. The available simulations can be determined by running
names(defineSimulations()). Otherwise, if the user would like to define a new
simulation on the fly, the argument simulations can take the form of a named list of
new functions (simulations).}

\item{nulls}{Optional. If not provided, defines the nulls as all of those in
defineNulls. If only a subset of those is desired, then nulls should take
the form of a character vector corresponding to named functions from defineNulls.
The available nulls can be determined by running names(defineNulls()). Otherwise,
if the user would like to define a new null on the fly, the argument nulls can take
the form of a named list of new functions (nulls).}

\item{metrics}{Optional. If not provided, defines the metrics as all of those in
defineBetaMetrics. If only a subset of those is desired, then metrics should take
the form of a character vector corresponding to named functions from defineBetaMetrics.
The available metrics can be determined by running names(defineBetaMetrics()).
If the user would like to define a new metric on the fly, the argument can take
the form of a named list of new functions (metrics).}
}
\value{
A list with two elements. The first is a list of data frames, with one for each
spatial simulation. These provide the observed beta metric scores for each spatial
simulation. The second level is a list of lists, one for each spatial simulation. Each
of these is a list of data frames. There is one data frame per null model, and it
summarizes the randomized metric scores for that null model for that spatial
simulation. Note that this is slightly different than the regular linker() function,
which does not output these raw metric scores (that function calculates SES and CI as
outputs).
}
\description{
This function wraps a number of wrapper functions into one big metric + null 
tester function. Only a single test is performed, with results saved into memory.
}
\details{
This function wraps a number of other wrapper functions into
one big beta metric + null performance tester function. Only a single test is run, 
with results saved into memory. To perform multiple complete tests, use the
multiLinker function, which saves results to file.
}
\examples{
system.time(test <- betaLinker(no.taxa=50, arena.length=300, mean.log.individuals=2, 
	length.parameter=5000, sd.parameter=50, max.distance=30, proportion.killed=0.2, 
competition.iterations=3, no.plots=15, plot.length=30,
randomizations=3, cores="seq", metrics=c("Pst", "Bst"),
nulls=c("richness", "frequency")))
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2016. Phylogenetic community
structure metrics and null models: a review with new methods and software.
Ecography DOI: 10.1111/ecog.02070
}
