% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkMetrics.R
\name{checkMetrics}
\alias{checkMetrics}
\title{Confirm that the metric functions are in suitable format}
\usage{
checkMetrics(x)
}
\arguments{
\item{x}{Optional named list of metric functions. Else, defines the metrics as those
defined in defineMetrics.}
}
\value{
A list of functions.
}
\description{
Utility function. Creates a list of functions, either those defined in defineMetrics
or a named list of metric functions.
}
\details{
A few quick checks to confirm the metric functions are input in suitable
format.
}
\examples{
checkMetrics(defineMetrics())
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

