% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNulls.R
\name{checkNulls}
\alias{checkNulls}
\title{Confirm that the null model functions are in suitable format}
\usage{
checkNulls(x)
}
\arguments{
\item{x}{Optional named list of null models. Else, defines the nulls as those
defined in defineNulls.}
}
\value{
A list of functions.
}
\description{
Utility function. Creates a list of null models, either those defined in defineNulls
or a named list of null model functions.
}
\details{
A few quick checks to confirm the null model functions are input in suitable
format.
}
\examples{
checkNulls(defineNulls())
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

