% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkSimulations.R
\name{checkSimulations}
\alias{checkSimulations}
\title{Confirm that the spatial simulation functions are in suitable format}
\usage{
checkSimulations(x)
}
\arguments{
\item{x}{Optional named list of spatial simulation functions. Else, defines the
spatial simulations as those in defineSimulations.}
}
\value{
A list of functions.
}
\description{
Utility function. Creates a list of spatial simulations, either those defined in
defineSimulations or a named list of simulation functions.
}
\details{
A few quick checks to confirm the spatial simulations are in suitable format.
}
\examples{
checkSimulations(defineSimulations())
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

