% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defineMetrics.R
\name{defineMetrics}
\alias{defineMetrics}
\title{Output all metrics as a list of named functions}
\usage{
defineMetrics()
}
\value{
A list of named functions
}
\description{
Creates a list of named functions, each of which accept a metrics.input object
}
\details{
All of the functions we calculated for our manuscript are included in this
function. To add additional functions, they can either be defined on the fly or, to
permanently include a new metric in all downstream simulations, it can be included
here. The function needs to be included with a name, and it must accept a metrics.input
as input. If the function needs additional elements not included in that input, then
the prepData function must also be revised.
}
\examples{
defineMetrics()
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

