% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMetrics.R
\name{defineMetrics}
\alias{defineMetrics}
\title{Output all metrics as a list of named functions}
\usage{
defineMetrics()
}
\value{
A list of named functions
}
\description{
Creates a list of named functions, each of which accept a metrics.input object
}
\details{
All of the functions we calculated for our manuscript are included in this
function. To add additional functions, they can either be defined on the fly or, to 
permanently include a new metric in all downstream simulations, it can be included
here. The function needs to be included with a name, and it must accept a metrics.input
as input. If the function needs additional elements not included in that input, then
the prepData function must also be revised. Currently MRD references the
non-abundance-weighted version of the function. This is simply so calcField knows
how to handle the results downstream. Reminder to modify that function if 
abundance-weighted MRD is to be implemented here.
}
\examples{
defineMetrics()
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2016. Phylogenetic community
structure metrics and null models: a review with new methods and software.
Ecography DOI: 10.1111/ecog.02070
}
