% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineNulls.R
\name{defineNulls}
\alias{defineNulls}
\title{Output all null models as a list of named functions}
\usage{
defineNulls()
}
\value{
A list of named functions
}
\description{
Creates a list of named functions, each of which accept a nulls.input object
}
\details{
All of the null models we calculated for our manuscript are included in this
function. To add additional null model functions, they can either be defined on the fly
or to permanently include a new model in all downstream simulations, it can be
included here. The function needs to be included with a name, and it must accept a
nulls.input object. If the function needs additional elements not included in that
input, then the prepNulls function must also be revised.
}
\examples{
defineNulls()
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2016. Phylogenetic community
structure metrics and null models: a review with new methods and software.
Ecography DOI: 10.1111/ecog.02070
}

