% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMRCA.R
\name{distMRCA}
\alias{distMRCA}
\title{Calculate distances to most recent common ancestors}
\usage{
distMRCA(samp, tree)
}
\arguments{
\item{samp}{A picante-style community data matrix with sites as rows, and
species as columns}

\item{tree}{An ape-style phylogeny.}
}
\value{
A vector of MRD values, calculated according to the method specified.
}
\description{
Given a picante-style community data matrix (sites are rows, species are columns), 
and a phylogeny, calculate the distances of sets of taxa to their MRCA.
}
\details{
A here define a new measure 
the mean branch length between a set of taxa and their most recent common ancestor
(MRCA). Large values of this new measure ("distMRCA") correspond to taxa with a
distant MRCA, while small values correspond to taxa with a more recent MRCA.
To calculate the original MRD, keep node.based as the default (TRUE). To access
distMRCA, set node.based to FALSE. In an ultrametric tree, the mean distance
between a set of taxa and a single ancestor is of course equal to the distance between
one of those taxa and the ancestor. However, in case an ultrametric tree is passed
to the function, I do indeed define it as the mean distance between all present taxa
and their MRCA. It will throw a warning if a non-ultrametric tree is passed along.
}
\examples{
#simulate tree with birth-death process
tree <- geiger::sim.bdtree(b=0.1, d=0, stop="taxa", n=50)

sim.abundances <- round(rlnorm(5000, meanlog=2, sdlog=1)) + 1

cdm <- simulateComm(tree, richness.vector=10:25, abundances=sim.abundances)

dists <- ape::cophenetic.phylo(tree)

results <- modifiedMPD(cdm, dists, abundance.weighted = "interspecific")
}
\references{
Miller, E. T. 2016. Random thoughts.

Kerr, J. T. and D. J. Currie. 1999. The relative importance of evolutionary and
environmental controls on broad-scale patterns of species richness in North America.
Ecoscience 6:329-337.
}

