% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nullPerformance}
\alias{nullPerformance}
\title{Summarize null model performance of a series of summarized simulation results}
\usage{
nullPerformance(summarized.results, simulations = "all", metrics = "all",
  concat.by = "both")
}
\arguments{
\item{summarized.results}{The results of a call to sesIndiv() or something similar.}

\item{simulations}{Default is "all". Alternatively, can supply a vector of simulation
names to summarize the results over.}

\item{metrics}{Default is "all". Alternatively, can supply a vector of metric
names to summarize the results over.}

\item{concat.by}{Default is "both". Alternatively, can supply either "quadrat" or
"richness".}
}
\value{
A data frame of summarized results
}
\description{
Flexible function that summarizes null model performance after reading in and testing
per-simulation results with a function like sesIndiv.
}
\details{
If an overall picture of null model performance is desired, this function can
provide it. It can also be used to summarize null model performance over a given set of
simulations, metrics, and concatenation options.
}
\examples{
#not run
#results <- readIn()
#summ <- sesIndiv(results)
#examp <- nullPerformance(summ)
}
\references{
Miller, Trisos and Farine.
}

