% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quadratOverall}
\alias{quadratOverall}
\title{Overall per simulation-null-metric quadrat test}
\usage{
quadratOverall(simulation.list, concat.by)
}
\arguments{
\item{simulation.list}{A summarized results list such as one output from
reduceResults(). See examples.}

\item{concat.by}{Whether randomizations were concatenated by richness, quadrat or both.}
}
\value{
A data frame summarizing the total number of quadrats tested and how many of
these deviated above (significantly overdispersed) or below (significantly clustered)
the 95% CI for each simulation, null, metric combination.
}
\description{
This function provides one of many ways of summarizing and considering simulation
results.
}
\details{
This function provides one way of summarizing and considering simulation
results. It takes as input a vector of 0s, 1s and 2s (corresponding to within, less,
and greater than the 95% CIs, respectively) for all quadrats
from a given simulation-null-metric combination, and determines how many quadrats
overall deviated beyond expectations. A number of utility functions are also defined
but not exported in this function.
}
\examples{
#not run
#results <- readIn()
#summ <- reduceResults(results, "both")
#test <- quadratOverall(summ$quadrat, "both")
}
\references{
Miller, Trisos and Farine.
}

