% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quadratPlotter.R
\name{quadratPlotter}
\alias{quadratPlotter}
\title{Plot simulated quadrats in arena}
\usage{
quadratPlotter(quadrat.bounds)
}
\arguments{
\item{quadrat.bounds}{Matrix of quadrat bounds}
}
\value{
Plotted quadrats
}
\description{
Given a matrix of quadrat bounds, plots the quadrats in an already plotted,
simulated arena.
}
\details{
Plots quadrats as defined by the supplied matrix, e.g. a call to
quadratPlacer. An active plot with the simulated arena needs to already be open,
see example.
}
\examples{
library(colorRamps)

tree <- sim.bdtree(b=0.1, d=0, stop="taxa", n=50)

temp <- evolveTraits(tree)

phydistmatrix <- cophenetic(temp[[1]])

#define a color for each species
cols <- blue2green2red(nrow(phydistmatrix))

#prep the data for the simulation
prepped <- prepSimulations(tree, arena.length=300, mean.log.individuals=2,
length.parameter=5000, sd.parameter=50, max.distance=20, proportion.killed=0.2,
competition.iterations=3)

positions <- filteringArena(prepped)

#plot the arena. don't close the window
plot(positions$arena$X, positions$arena$Y, pch=20, cex=0.5, xlim=c(0,300), ylim=c(0,300),
col=cols[positions$arena$individuals])

bounds <- quadratPlacer(no.quadrats=10, arena.length=300,
	quadrat.length=50)$quadrat.bounds

quadratPlotter(bounds)
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

