% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readIn.R
\name{readIn}
\alias{readIn}
\title{Read in the results of multiple metric/null/simulation tests}
\usage{
readIn(working.directory)
}
\arguments{
\item{working.directory}{Optional character string specifying the working directory.
If missing, the current working directory will be used.}
}
\value{
A list of simulation results. Each element in the list consists of a single
output from multiLinker().
}
\description{
This function is used to read all .RDS files from the working directory into a
single list. This list is then summarized with additional functions.
}
\details{
This function reads all .RDS files from the working directory (or another
specified directory) into a single list.
}
\examples{
#not run
#results <- readIn()
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

