% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reduceResults}
\alias{reduceResults}
\title{Reduce results from multiLinker into a manageable format}
\usage{
reduceResults(results.list, concat.by)
}
\arguments{
\item{results.list}{The results of a call to readIn()}

\item{concat.by}{Whether randomizations were concatenated by richness, quadrat or both}
}
\value{
A list of data frames.
}
\description{
Calling multiLinker creates .RDS files, one per iteration. This function will
combine these results into a more manageable format.
}
\details{
Given a list of results readIn() from multiLinker, this function will reduce
those results into a manageable format like that expected for calls to quadratOverall
and sesOverall.
}
\references{
Miller, Trisos and Farine.
}

