% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceResults.R
\name{reduceResults}
\alias{reduceResults}
\title{Reduce results from multiLinker into a manageable format}
\usage{
reduceResults(results.list, concat.by)
}
\arguments{
\item{results.list}{The results of a call to readIn()}

\item{concat.by}{Whether randomizations were concatenated by richness, plot or both}
}
\value{
A list of data frames.
}
\description{
Calling multiLinker creates .RDS files, one per iteration. This function will
combine these results into a more manageable format.
}
\details{
Given a list of results readIn() from multiLinker, this function will reduce
those results into a manageable format like that expected for calls to plotOverall
and sesOverall.
}
\examples{
#not run
#results <- readIn()
#summ <- reduceResults(results, "both")
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

