% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scaler.R
\name{scaler}
\alias{scaler}
\title{Scale output of evolveTraits to arena size}
\usage{
scaler(input.traits, min.arena, max.arena)
}
\arguments{
\item{input.traits}{Second element of the results of a call to evolveTraits()}

\item{min.arena}{Minimum size of arena, e.g. 0}

\item{max.arena}{Maximum size of arena}
}
\value{
A scaled and named dataframe of species traits
}
\description{
Given a matrix of two traits, and the minimum and maximum extent of the desired arena,
will return a data frame of species' traits scaled to the arena size.
}
\details{
Scales a matrix of species' traits to a desired mininimum-maximum range.
Intended for use in a spatially explicit scenario with two traits, but could easily
be co-opted.
}
\examples{
tree <- geiger::sim.bdtree(b=0.1, d=0, stop="taxa", n=50)

temp <- evolveTraits(tree)

scaled <- scaler(temp[[2]], min.arena=0, max.arena=300)
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

