% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sesSingle}
\alias{sesSingle}
\title{Summary of results from a single iteration}
\usage{
sesSingle(single.iteration)
}
\arguments{
\item{single.iteration}{The results of a single iteration from multiLinker.}
}
\description{
Use Wilcoxon signed rank test to determine whether quadrats from a SINGLE iteration
differ from expectations
}
\details{
This function uses a Wilcoxon signed rank test to determine whether the
quadrats from a spatial simulation/null/metric from a SINGLE iteration differ from
expectations. Assuming there are three spatial simulations named random, filtering, and
competition, this function will use two.sided, lesser and greater Wilcoxon tests,
respectively.
}
\examples{
#not run
#results <- readIn()
#summ <- sesSingle(results[[1]])
}
\references{
Miller, Trisos and Farine.
}

