% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wilcoWrapLApply.R
\name{wilcoWrapLApply}
\alias{wilcoWrapLApply}
\title{lapply wrapper for Wilcoxon-signed rank tests}
\usage{
wilcoWrapLApply(null.list, alternative)
}
\arguments{
\item{null.list}{A list of dataframes, one per null model, of observed metric scores.}

\item{alternative}{Optional alternative hypothesis. Default is "two-sided". Use
"greater" for competition; "less" for habitat filtering.}
}
\value{
A dataframe, with one row for each metric. The first column is the mean of the
vector of metric values, the second is the p.value of whether it differs from mu=0,
and the third is the name of the metric.
}
\description{
Just a utility function. lapplies wilco.test functions over a list of dataframes.
}
\details{
lapplies wilcoWrapApply over a list of dataframes.
}
\examples{
a <- rnorm(n=100)
b <- rnorm(n=100, mean=100)
ex <- data.frame(a, b)
test <- list("ex1"=ex, "ex2"=ex)
wilcoWrapLApply(test, alternative="two.sided")
}
\references{
Miller, E. T., D. R. Farine, and C. H. Trisos. 2015. Phylogenetic community
structure metrics and null models: a review with new methods and software.
bioRxiv 025726.
}

